+++ src/main/java/wooteco/subway/admin/controller/ServicePageController.java
@@ -1,0 +1,18 @@
+package wooteco.subway.admin.controller;
+
+import org.springframework.http.MediaType;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class ServicePageController {
+    @GetMapping(value = "/", produces = MediaType.TEXT_HTML_VALUE)
+    public String index() {
+        return "service/index";
+    }
+
+    @GetMapping(value = "/search", produces = MediaType.TEXT_HTML_VALUE)
+    public String searchPage() {
+        return "service/search";
+    }
+}
