+++ src/main/java/wooteco/subway/admin/service/StationService.java
@@ -1,0 +1,20 @@
+package wooteco.subway.admin.service;
+
+import org.springframework.stereotype.Service;
+import wooteco.subway.admin.domain.Station;
+import wooteco.subway.admin.repository.StationRepository;
+
+import java.util.List;
+
+@Service
+public class StationService {
+    private StationRepository stationRepository;
+
+    public StationService(StationRepository stationRepository) {
+        this.stationRepository = stationRepository;
+    }
+
+    public List<Station> findAll() {
+        return stationRepository.findAll();
+    }
+}
